Engine - added vec3 and mat3

+++ Engine/Shapes.h
@@ -8,7 +8,7 @@ public:
 	static std::vector<Vec2<float>> makeStar(float outerRadius, float innerRadius, int nFlares = 5)
 	{
 		std::vector<Vec2<float>> star;
-		star.reserve(nFlares * 2);
+		star.reserve((size_t)nFlares * 2);
 		const float dTheta = 2.0f * 3.14159f / float(nFlares * 2);
 		for (int i = 0; i < nFlares * 2; i++)
 		{
